requirements.txt
anyio==4.9.0
click==8.1.8
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.95.2
gunicorn==20.1.0
h11==0.16.0
idna==3.10
lxml==6.0.0
openpyxl==3.1.2
packaging==25.0
Pillow==10.0.0
pydantic==1.10.22
PyMuPDF==1.22.5
pytesseract==0.3.10
python-docx==0.8.11
python-multipart==0.0.6
python-pptx==0.6.21
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.1
uvicorn==0.22.0
xlsxwriter==3.2.5
